chore: move two commands from chain cli to eth cli
