+++ app/submodule/chain/chaininfo_api.go
@@ -562,7 +562,7 @@ func (cia *chainInfoAPI) StateWaitMsg(ctx context.Context, mCid cid.Cid, confide
 			t, err := cia.getReturnType(ctx, vmsg.To, vmsg.Method)
 			if err != nil {
 				if errors.Is(err, ErrMetadataNotFound) {
-					// This is not nececessary an error -- EVM methods (and in the future native actors) may
+					// This is not necessarily an error -- EVM methods (and in the future native actors) may
 					//  return just bytes, and in the not so distant future we'll have native wasm actors
 					//  that are by definition not in the registry.
 					// So in this case, log a debug message and retun the raw bytes.
@@ -864,7 +864,7 @@ func (cia *chainInfoAPI) StateReplay(ctx context.Context, tsk types.TipSetKey, m
 // ChainGetEvents returns the events under an event AMT root CID.
 func (cia *chainInfoAPI) ChainGetEvents(ctx context.Context, root cid.Cid) ([]types.Event, error) {
 	store := cbor.NewCborStore(cia.chain.ChainReader.Blockstore())
-	evtArr, err := amt4.LoadAMT(ctx, store, root, amt4.UseTreeBitWidth(5))
+	evtArr, err := amt4.LoadAMT(ctx, store, root, amt4.UseTreeBitWidth(types.EventAMTBitwidth))
 	if err != nil {
 		return nil, fmt.Errorf("load events amt: %w", err)
 	}
