+++ cmd/state.go
@@ -359,7 +359,7 @@ var stateGetActorCmd = &cmds.Command{
 		writer.Printf("Nonce:\t\t%d\n", a.Nonce)
 		writer.Printf("Code:\t\t%s (%s)\n", a.Code, strtype)
 		writer.Printf("Head:\t\t%s\n", a.Head)
-		writer.Printf("Predictable address:\t\t%s\n", a.Address)
+		writer.Printf("Delegated address:\t\t%s\n", a.Address)
 
 		return re.Emit(buf)
 	},
