+++ venus-shared/types/message_receipt.go
@@ -28,7 +28,7 @@ type MessageReceipt struct {
 	Return   []byte
 	GasUsed  int64
 
-	EventsRoot *cid.Cid // Root of Event AMT
+	EventsRoot *cid.Cid // Root of Event AMT with bitwidth = EventAMTBitwidth
 }
 
 // NewMessageReceiptV0 creates a new pre FIP-0049 receipt with no capability to
